+++ backend/connector-integration/src/connectors.rs
@@ -61,3 +61,6 @@ pub use self::bluecode::Bluecode;
 
 pub mod cryptopay;
 pub use self::cryptopay::Cryptopay;
+
+pub mod helcim;
+pub use self::helcim::Helcim;